tee_shm_allocate(): use size_t for size and check for overflow
Fixes failure of xtest 7002 on Foundation_v8 (PLATFORM=vexpress-fvp).

xtest 7002 calls TEEC_AllocateSharedMemory() with a requested size of
0xFFFFFFFE and expects an out-of-memory error. Prior to this commit,
the size after page_size alignment would wrap to 0 and no error would be
returned.

Tested on PLATFORM vexpress-fvp and vexpress-qemu_virt.

Tested-by: Pascal Brand <pascal.brand@linaro.org> (STM platform)

+++ generic/tee_mem.h
@@ -42,7 +42,7 @@ unsigned long tee_shm_pool_alloc(struct device *dev,
 
 
 void tee_shm_pool_free(struct device *dev, struct shm_pool *pool,
-		unsigned long paddr, uint32_t *size);
+		unsigned long paddr, size_t *size);
 
 
 bool tee_shm_pool_incref(struct device *dev, struct shm_pool *pool,
